requirements.txt
langchain~=0.3.25
langchain_openai~=0.3.18
python-dotenv~=1.1.0
pyyaml~=6.0.2
streamlit~=1.50.0
fastapi~=0.117.1
uvicorn~=0.37.0
